Initialize square's words to be None
This makes it possible to call is_start(), is_end(), and
displayed_clue_number() on "free-floating squares", i.e.
squares with no words, without it blowing up with a KeyError:

>>> import xword
>>> square = xword.Square(0, 0, 'A')
>>> square.is_start(xword.Direction.ACROSS)
False
>>> square.displayed_clue_number() is None
True

Although in practice these methods will only be called on squares
with words.

I'm not sure if it's a good idea. I guess it's somewhat similar to

>>> from lxml import etree
>>> foo = etree.Element('foo')
>>> foo.getparent() is None
True